Heatmap. Use raw frequencies rather then statisically altered (enhanced) frequencies
